Add Streamlit debugging configuration and update data extraction functions to include project_id
