+++ front_end/data_extraction/residential_extractor.py
@@ -21,5 +21,5 @@ data_units = [
 
 data = dict(zip(data_names, zip(data_types, data_units)))
 
-def extract(models, client):
-    return data_extractor.extract(data, model_name, models, client)
+def extract(models, client, project_id):
+    return data_extractor.extract(data, model_name, models, client, project_id)
